change int types to long
The des api sometimes returned huge values
and the int data type couldn't hold them.

+++ Net.Pokeshot.JiveSdk/Models/Dto/JiveDEAActivityInstance.cs
@@ -8,11 +8,11 @@ namespace Net.Pokeshot.JiveSdk.Models.Dto
     public class JiveDEAActivityInstance
     {
         public string name { get; set; }
-        public int timestamp { get; set; }
+        public long timestamp { get; set; }
         public int seqId { get; set; }
         public string uuid { get; set; }
         public JiveDEAContext context { get; set; }
-        public int actorID { get; set; }
+        public long actorID { get; set; }
         public int actorType { get; set; }
         public string activityType { get; set; }
         public int actionObjectId { get; set; }
